switch local route binding to use netroute
