+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/libp2p/go-libp2p-core v0.5.1
 	github.com/libp2p/go-libp2p-tls v0.1.3
 	github.com/libp2p/go-maddr-filter v0.0.5
+	github.com/libp2p/go-netroute v0.1.2
 	github.com/lucas-clemente/quic-go v0.15.2
 	github.com/minio/sha256-simd v0.1.1
 	github.com/multiformats/go-multiaddr v0.2.1
@@ -14,7 +15,5 @@ require (
 	github.com/multiformats/go-multiaddr-net v0.1.4
 	github.com/onsi/ginkgo v1.12.0
 	github.com/onsi/gomega v1.9.0
-	github.com/vishvananda/netlink v1.1.0
 	golang.org/x/crypto v0.0.0-20200221231518-2aa609cf4a9d
-	golang.org/x/sys v0.0.0-20200223170610-d5e6a3e2c0ae
 )
