+++ libp2pquic_suite_test.go
@@ -32,7 +32,7 @@ var (
 func isGarbageCollectorRunning() bool {
 	var b bytes.Buffer
 	pprof.Lookup("goroutine").WriteTo(&b, 1)
-	return strings.Contains(b.String(), "go-libp2p-quic-transport.(*reuseBase).runGarbageCollector")
+	return strings.Contains(b.String(), "go-libp2p-quic-transport.(*reuse).runGarbageCollector")
 }
 
 var _ = BeforeEach(func() {
