+++ reuse_linux_test.go
@@ -13,9 +13,7 @@ var _ = Describe("Reuse (on Linux)", func() {
 	var reuse *reuse
 
 	BeforeEach(func() {
-		var err error
-		reuse, err = newReuse(nil)
-		Expect(err).ToNot(HaveOccurred())
+		reuse = newReuse(nil)
 	})
 
 	Context("creating and reusing connections", func() {
